# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/CoordinatorDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class CoordinatorDto
    {
        /// <summary>
        /// Initializes a new instance of the CoordinatorDto class.
        /// </summary>
        public CoordinatorDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the CoordinatorDto class.
        /// </summary>
        public CoordinatorDto(System.Guid? id = default(System.Guid?), string name = default(string), string email = default(string), string website = default(string), string cellPhone = default(string), string phoneNumber = default(string))
        {
            Id = id;
            Name = name;
            Email = email;
            Website = website;
            CellPhone = cellPhone;
            PhoneNumber = phoneNumber;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public System.Guid? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "email")]
        public string Email { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "website")]
        public string Website { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "cellPhone")]
        public string CellPhone { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "phoneNumber")]
        public string PhoneNumber { get; set; }

    }
}