# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/EventDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public partial class EventDto
    {
        /// <summary>
        /// Initializes a new instance of the EventDto class.
        /// </summary>
        public EventDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the EventDto class.
        /// </summary>
        public EventDto(System.Guid? eventId = default(System.Guid?), string name = default(string), string description = default(string), PriceDto price = default(PriceDto), LocationDto location = default(LocationDto), CoordinatorDto coordinator = default(CoordinatorDto), IList<ImageDto> images = default(IList<ImageDto>), System.DateTime? fromDate = default(System.DateTime?), System.DateTime? toDate = default(System.DateTime?), System.DateTime? createdDate = default(System.DateTime?), System.DateTime? modifiedDate = default(System.DateTime?))
        {
            EventId = eventId;
            Name = name;
            Description = description;
            Price = price;
            Location = location;
            Coordinator = coordinator;
            Images = images;
            FromDate = fromDate;
            ToDate = toDate;
            CreatedDate = createdDate;
            ModifiedDate = modifiedDate;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "eventId")]
        public System.Guid? EventId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "description")]
        public string Description { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "price")]
        public PriceDto Price { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "location")]
        public LocationDto Location { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "coordinator")]
        public CoordinatorDto Coordinator { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "images")]
        public IList<ImageDto> Images { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "fromDate")]
        public System.DateTime? FromDate { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "toDate")]
        public System.DateTime? ToDate { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "createdDate")]
        public System.DateTime? CreatedDate { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "modifiedDate")]
        public System.DateTime? ModifiedDate { get; set; }

    }
}