# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/GeoCoordinateDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class GeoCoordinateDto
    {
        /// <summary>
        /// Initializes a new instance of the GeoCoordinateDto class.
        /// </summary>
        public GeoCoordinateDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the GeoCoordinateDto class.
        /// </summary>
        public GeoCoordinateDto(System.Guid? id = default(System.Guid?), double? latitude = default(double?), double? longitude = default(double?))
        {
            Id = id;
            Latitude = latitude;
            Longitude = longitude;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public System.Guid? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "latitude")]
        public double? Latitude { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "longitude")]
        public double? Longitude { get; set; }

    }
}