# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/HolidayTypeDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class HolidayTypeDto
    {
        /// <summary>
        /// Initializes a new instance of the HolidayTypeDto class.
        /// </summary>
        public HolidayTypeDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the HolidayTypeDto class.
        /// </summary>
        public HolidayTypeDto(System.Guid? id = default(System.Guid?), string name = default(string))
        {
            Id = id;
            Name = name;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public System.Guid? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

    }
}