# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/ImageDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class ImageDto
    {
        /// <summary>
        /// Initializes a new instance of the ImageDto class.
        /// </summary>
        public ImageDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the ImageDto class.
        /// </summary>
        public ImageDto(System.Guid? id = default(System.Guid?), string name = default(string), string base64 = default(string), string sourceURL = default(string))
        {
            Id = id;
            Name = name;
            Base64 = base64;
            SourceURL = sourceURL;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public System.Guid? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "base64")]
        public string Base64 { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "sourceURL")]
        public string SourceURL { get; set; }

    }
}