# Adventure.Mobile/Adventure.Client.API/Autogenerated/Models/PriceDto.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Adventure.Client.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class PriceDto
    {
        /// <summary>
        /// Initializes a new instance of the PriceDto class.
        /// </summary>
        public PriceDto()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the PriceDto class.
        /// </summary>
        public PriceDto(System.Guid? id = default(System.Guid?), double? kidsPrice = default(double?), double? adultsPrice = default(double?), double? seniorCitizenPrice = default(double?))
        {
            Id = id;
            KidsPrice = kidsPrice;
            AdultsPrice = adultsPrice;
            SeniorCitizenPrice = seniorCitizenPrice;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public System.Guid? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "kidsPrice")]
        public double? KidsPrice { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "adultsPrice")]
        public double? AdultsPrice { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "seniorCitizenPrice")]
        public double? SeniorCitizenPrice { get; set; }

    }
}